some avatar and small things made

+++ app/Http/Controllers/Provider/CustomerController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Provider;
 
 use App\Http\Controllers\Controller;
 use App\Models\User;
+use App\Services\Sms;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
@@ -33,18 +34,35 @@ class CustomerController extends Controller
             'last_name' => 'required',
             'gender' => 'required',
             'type' => 'required',
-            'email' => 'required',
             'phone' => 'required',
-            'password' => 'required',
-            'password_confirmation' => 'required',
         ]);
 
         $provider = auth()->user()->provider;
 
         if (!User::where('phone', '=', $attributes['phone'])->exists()) {
-            $user = User::create($attributes);
+            $password = $this->generate_password();
+            $user = User::create([
+                'first_name' => $attributes['first_name'],
+                'last_name' => $attributes['last_name'],
+                'gender' => $attributes['gender'],
+                'type' => $attributes['type'],
+                'phone' => $attributes['phone'],
+                'password' => $password
+            ]);
+            $sms = new Sms($attributes['phone'], 'newCustomer', [
+                $user->first_name . ' ' . $user->last_name,
+                $provider->title,
+                $attributes['phone'],
+                $password
+            ]);
+            $sms->send();
         } else {
             $user = User::where('phone', '=', $attributes['phone'])->first();
+            $sms = new Sms($attributes['phone'], 'addCustomer', [
+                $user->first_name . ' ' . $user->last_name,
+                $provider->title
+            ]);
+            $sms->send();
         }
 
         $provider->customers()->attach($user->id);
@@ -54,8 +72,14 @@ class CustomerController extends Controller
 
     public function edit(User $user)
     {
+        $canEdit = true;
+
+        if ($user->hasVerifiedPhone())
+            $canEdit = false;
+
         return Inertia::render('Provider/User/Edit', [
-            'current_user' => $user
+            'current_user' => $user,
+            'canEdit' => $canEdit
         ]);
     }
 
@@ -65,15 +89,10 @@ class CustomerController extends Controller
             'first_name' => 'required',
             'last_name' => 'required',
             'gender' => 'required',
-            'email' => 'required',
             'phone' => 'required',
-            'password' => '',
-            'password_confirmation' => '',
         ]);
-
-        $attributes = array_filter($attributes);
-
-        $user->update($attributes);
+        if ($user->hasVerifiedPhone())
+            $user->update($attributes);
 
         return redirect('/provider/users/' . $user->id . '/edit');
     }
@@ -82,9 +101,24 @@ class CustomerController extends Controller
     {
         $provider = auth()->user()->provider;
 
-        $provider->customers()->detach($user->id);
+        if ($user->hasVerifiedPhone())
+            $provider->customers()->detach($user->id);
 
         return redirect('/provider/users');
+    }
+
+    public function generate_password($length = 8): string
+    {
+        $chars =  'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz'.
+            '0123456789=!@#$%&*+?';
+
+        $str = '';
+        $max = strlen($chars) - 1;
+
+        for ($i=0; $i < $length; $i++)
+            $str .= $chars[random_int(0, $max)];
+
+        return $str;
     }
 
 }
